bot/requirements.txt
python-telegram-bot>=22.0
google-genai>=1.5.0
supabase>=2.10.0
python-dotenv==1.0.1
pdfplumber==0.11.9
pdf2image==1.17.0
Pillow==12.1.1
google-api-python-client>=2.122.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
